Split project, organization, context to seperate app
