+++ context/serializers.py
@@ -1,0 +1,15 @@
+from rest_framework import serializers
+
+from .models import Context, Module
+
+
+class ContextSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Context
+        fields = "__all__"
+
+
+class ModuleSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Module
+        fields = "__all__"
