+++ project/apps.py
@@ -1,0 +1,9 @@
+from django.apps import AppConfig
+
+
+class ProjectConfig(AppConfig):
+    default_auto_field = "django.db.models.BigAutoField"
+    name = "project"
+
+    def ready(self):
+        from project import signals
