+++ project/models.py
@@ -1,0 +1,76 @@
+from django.conf import settings
+from django.db import models
+from ordered_model.models import OrderedModel
+
+from neatplus.models import TimeStampedModel, UserStampedModel
+
+
+class Project(TimeStampedModel, UserStampedModel, OrderedModel):
+    class VisibilityChoice(models.TextChoices):
+        PUBLIC = "public"
+        PUBLIC_WIITHIN_ORGANIZATION = "public_within_organization"
+        PRIVATE = "private"
+
+    class StatusChoice(models.TextChoices):
+        PENDING = "pending"
+        ACCEPTED = "accepted"
+        REJECTED = "rejected"
+
+    title = models.CharField(max_length=255)
+    description = models.TextField()
+    organization = models.ForeignKey(
+        "organization.Organization", on_delete=models.CASCADE, related_name="projects"
+    )
+    visibility = models.CharField(max_length=26, choices=VisibilityChoice.choices)
+    users = models.ManyToManyField(
+        settings.AUTH_USER_MODEL,
+        related_name="projects",
+        through="ProjectUser",
+        through_fields=("project", "user"),
+    )
+    status = models.CharField(
+        max_length=8,
+        choices=StatusChoice.choices,
+        default=StatusChoice.PENDING,
+        editable=False,
+    )
+    context = models.ForeignKey(
+        "context.Context", on_delete=models.CASCADE, related_name="projects"
+    )
+
+    class Meta(OrderedModel.Meta):
+        pass
+
+    def __str__(self):
+        return self.title
+
+    def save(self, *args, **kwargs):
+        if self.pk:
+            cls = self.__class__
+            old = cls.objects.get(pk=self.pk)
+            changed_fields = []
+            for field in cls._meta.get_fields():
+                field_name = field.name
+                try:
+                    if getattr(old, field_name) != getattr(self, field_name):
+                        changed_fields.append(field_name)
+                except Exception as e:
+                    pass
+            kwargs["update_fields"] = changed_fields
+            if "organization" in kwargs["update_fields"]:
+                self.status = "pending"
+        super().save(*args, **kwargs)
+
+
+class ProjectUser(UserStampedModel, TimeStampedModel):
+    class SurveyPermissionChoice(models.TextChoices):
+        WRITE = "write"
+        READ_ONLY = "read_only"
+
+    project = models.ForeignKey("Project", on_delete=models.CASCADE)
+    user = models.ForeignKey(settings.AUTH_USER_MODEL, on_delete=models.CASCADE)
+    permission = models.CharField(
+        max_length=9,
+        choices=SurveyPermissionChoice.choices,
+        default=SurveyPermissionChoice.READ_ONLY,
+    )
