+++ survey/admin.py
@@ -4,34 +4,7 @@ from ordered_model.admin import OrderedModelAdmin
 
 from neatplus.admin import UserStampedModelAdmin
 
-from .models import Answer, Context, Module, Question, QuestionCategory
-
-
-@admin.register(Context)
-class ContextAdmin(
-    UserStampedModelAdmin,
-    TranslationAdmin,
-    OrderedModelAdmin,
-):
-    list_display = ("code", "title", "move_up_down_links")
-    search_fields = (
-        "code",
-        "title",
-    )
-
-
-@admin.register(Module)
-class ModuleAdmin(
-    UserStampedModelAdmin,
-    TranslationAdmin,
-    OrderedModelAdmin,
-):
-    list_display = ("code", "title", "context", "move_up_down_links")
-    autocomplete_fields = ("context",)
-    search_fields = (
-        "code",
-        "title",
-    )
+from .models import Answer, Question, QuestionCategory
 
 
 @admin.register(QuestionCategory)
