+++ survey/models.py
@@ -4,33 +4,10 @@ from ordered_model.models import OrderedModel
 from neatplus.models import CodeModel, TimeStampedModel, UserStampedModel
 
 
-class Context(CodeModel, UserStampedModel, TimeStampedModel, OrderedModel):
-    title = models.CharField(max_length=255)
-
-    def __str__(self):
-        return self.code + " " + self.title
-
-    class Meta(OrderedModel.Meta):
-        pass
-
-
-class Module(CodeModel, UserStampedModel, TimeStampedModel, OrderedModel):
-    title = models.CharField(max_length=255)
-    context = models.ForeignKey(
-        "Context", on_delete=models.PROTECT, related_name="modules"
-    )
-
-    def __str__(self):
-        return self.code + " " + self.title
-
-    class Meta(OrderedModel.Meta):
-        pass
-
-
 class QuestionCategory(CodeModel, UserStampedModel, TimeStampedModel, OrderedModel):
     title = models.CharField(max_length=255)
     module = models.ForeignKey(
-        "Module", on_delete=models.PROTECT, related_name="categories"
+        "context.Module", on_delete=models.PROTECT, related_name="categories"
     )
 
     def __str__(self):
