+++ survey/translation.py
@@ -1,16 +1,6 @@
 from modeltranslation import translator
 
-from .models import Answer, Context, Module, Question, QuestionCategory
-
-
-@translator.register(Context)
-class ContextTranslationOptions(translator.TranslationOptions):
-    fields = ("title",)
-
-
-@translator.register(Module)
-class ModuleTranslationOptions(translator.TranslationOptions):
-    fields = ("title",)
+from .models import Answer, Question, QuestionCategory
 
 
 @translator.register(QuestionCategory)
